// frontend/rodia-monorepo/apps/mobile/src/shared/api/generated/schemas/announcementResponse.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface AnnouncementResponse {
  announcementId?: number;
  adminId?: number;
  title?: string;
  content?: string;
  isPinned?: boolean;
  publishedAt?: string;
  createdAt?: string;
  updatedAt?: string;
}